Made typewriter externally usable; implemented src/lib.rs with type_print function.
